+++ tools/rescue/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod admin_script_builder;
-pub mod mission;
+pub mod rescue_tx;
 pub mod framework_payload;
 pub mod diem_db_bootstrapper;
